Added Startup Scene with Data Manager and Event Manager
The Data Manager is a singleton class that stores persistent information.

The Event Manager is a class that handles events between scripts.

The Startup Scene is meant to be the first scene loaded in the game.
Its main functionality is to allow Data Manager to initialize all persistent
data to default values, if not yet already.  Once all data has been
initialized, the Data Manager triggers an event to load the next
scene (event code is ready but the new scene has not yet been created).

The default SampleScene has been removed.