Added MainMenuScene with dummy buttons
The MainMenuScene is now loaded from the StartupScene instead
of the TestScene.

Dummy buttons are placed in the MainMenuScene for now.  The
back button however has been programmed to display an Exit Panel
which prompts the user for confirmation to exit the app.  The
exit panel is fully functional as of this commit.

There were also some cleanups done for the TestUI script, which
were the renaming of some methods.