Added touch and square movement handling for LevelScene
The LevelLogic now calculates movements and tile positions by
using world map conventions which is by x and y, rather than by
the Level layout convention of row and column.  A translation
is therefore performed when loading the Level layout.

Also, some fixes are performed on the StartupScene and MapMenuScene.