Added star counter in LevelScene
The number of stars will be awarded depending on the number of
moves that the user obtained.

If the user completed the level with the same number of hint
length of map, 3 stars is awarded.  2 if they need and extra
step and 1 if otherwise.

The largest number of stars awarded is written into the DataManager
and is reloaded on the next run.

The Level GameObject in LevelScene is now renamed to Map to avoid
confusion with the Level scripts.