Added pause capability and reworked interactability of Control buttons in LevelScene
A Pause Button is added into the Control Panel which when pressed
will open a Pause Panel.  The Pause Panel currently has a single
Resume Button which when pressed closes the Pause Panel.

The interactability of Control Buttons are now turned off the
moment a touch is rather then when a swipe is detected.