Added ability to continue last played level from MainMenuScene
Three new data are added into DataManager: LastColor, LastAlphabet
and LastMap which is written with MenuColor, MenuAlphabet and
MenuMap when a LevelScene is loaded.

In the MainMenuScene, when the Continue Button is pressed, the
last level based on the three new data is loaded.

If the user is playing the game for the first time, the first
level in the entire game is loaded instead.