Added Ad Button on Control Panel in LevelScene
The Ad Button is displayed over the Hint Button should the user
no longer have any hints.  Pressing the Button will play a rewarded
ad which will reward the user with 1 hint.

The Rewatch Button in the HintAbort panel is now removed.  This is
because there isn't much necessity for this button as the player
can simply click the Ad button again after closing the HintAbort
panel.