Added AdLoad panel in LevelScene
The AdLoad panel will be displayed before an ad is load.  To do
this new touchStates are added to run the AdLoad Panel.  In those
states, the program loops for a cetain amount of time until either
an ad is successfully loaded, or if the maximum allocated time is
up.