Added CloudOnce and configured single HighScore achievement for Android
CloudOnce will be initialized in the Startup and StartupLogic will wait
for the initialization to complete before loading the MainMenuScene.
Event Manager is also phased out and the events and delegates are provided
by the respective classes.

There is currently no way to view the leaderboard from the game, although
the number of stars are already updated after completing each level.